pyproject.toml
[project]
name = "benj-office-hero"
version = "0.1.0"
description = "Office Hero web application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "httpx>=0.24",
    "click>=8.0",
    "sqlalchemy>=2.0",
    "alembic>=1.10",
    "asyncpg>=0.27",
    "fastapi>=0.104",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "slowapi>=0.1",
    "structlog>=24.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.24",
    "aiosqlite>=0.19",
    "pytest-cov>=4.1",
    "pre-commit>=4.0",
    "ruff>=0.8",
    "black>=24.0",
    "bandit>=1.8",
    "hypothesis>=6.88",
    "pytest-xdist>=3.5",
    "pytest-timeout>=2.2",
    "httpx>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=0",
    # Raise to 50+ as the project grows:
    # "--cov-fail-under=50",
    "--timeout=30",
]
markers = [
    "e2e: end-to-end browser tests",
    "slow: slow integration tests",
    "unit: fast unit tests",
    "integration: integration tests",
]

[tool.coverage.run]
branch = true
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
precision = 2

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
